Work on setup wizard
- Adding some steps for the setup
- Adding some icons for the plugins

Adding icons for the setup.

Styling the plugin buttons and copy the plugins when validating the PluginsStep

- Adding an "install all" check box for plugins
- InstallationLocationStep

Working on the general arrangement of the wizard

The current step is displayed in bold in the steps' list

Styling the setup view

Application setup refactor

Small refactor to handle exceptions during steps.

Use custom exception for handling error during installation.

Adding a finish step and make the cancellation of the setup centralized.

Trying to better package the application.

Improve packaging

WIP on a better packaging

WIP on a better packaging

Packaging now includes setup.

Documentation

Correct usage of the distribution plugin for the packaging.

Refactoring of the artifact packaging.

Documentation

Refactoring of the artifacts packaging.

Trying to package correctly

Packaging seems to be complete

Working on embedding all artifacts within the setup.

Working on embedding all artifacts within the setup.

Packaging and gradle build scripts improvements.

Include documentations in the installed location.